Add run-task-capacity-provider-strategy input
When you use cluster auto scaling, you must specify `capacityProviderStrategy` and not `launchType`. Unfortunately this is not possible to configure when running a one-off task with this action.

This adds a new input `run-task-capacity-provider-strategy` which takes a JSON array of capacity provider strategy items. When this is specified, `run-task-launch-type` will not be applied.